fm-thumbnail-loader: Dirty hack to work around exit deadlock
run_thumbnailer deadlocks in g_cond_wait when the program exits because
sometimes _pid_watcher is never exexuted even though the child has
exited. Poll the fields of GChildWatchSource directly to work around
this. It's not safe as this is an internal structure that isn't exposed
in any header, but it works.